Fixed issue introduced with JQ3.6.0 where keydown on s2 element did not select matching option
JQ 3.6.0 introduced a fix for a bug related to setting focus on an element within another elements on focus event. This fix caused an issue in select2 where when a select2 element was focussed and a user typed, rather than selecting a matching option, the dropdown wasn’t filtered.

This change swaps the jQuery trigger focus for a native .focus() call. Tested in JQ 3.4.1 and 3.6.0